fixing 1.install

+++ scripts/1.install_monan.bash
@@ -153,13 +153,12 @@ rm -fr $MONANDIR/src/core_atmosphere/inc $MONANDIR/src/core_init_atmosphere/inc
 
 #CR: TODO: maybe later move this make script to main scripts directory.
 
-if [ ! -f "${DIR_SCRIPTS}/monan.sif" ]
-then
-    echo ""
+#downloading monan.SIF 
+if [[ "${MACHINE}" == *SIF && ! -f "${DIR_SCRIPTS}/monan.sif" ]]; then
     echo -e "${GREEN}==>${NC} downloading container monan.sif from dataserver......\n"
-
     wget -O "${DIR_SCRIPTS}/monan.sif" https://dataserver.cptec.inpe.br/dataserver_dimnt/monan/MONAN-Model/monan.sif
 fi
+
 
 echo ""
 echo -e  "${GREEN}==>${NC} Making compile script...\n"
